requirements.txt
alembic
appdirs
asgiref
black
certifi
click
colorama
distlib
fastapi[all]
uvicorn[standard]
FastAPI-SQLAlchemy
filelock
greenlet
gunicorn
h11
hypercorn
Mako
MarkupSafe
mypy-extensions
mysql-connector-python
passlib[bcrypt]
pathspec
pipenv
psycopg2
pydantic
pyjwt
pymysql
python-dateutil
python-dotenv
python-editor
python-jose[cryptography]
python-multipart
regex
six
black
SQLAlchemy
starlette
toml
typing-extensions
virtualenv
virtualenv-clone
Werkzeug
websockets
oauthlib